requirements.txt
fastapi[all]
sqlalchemy
passlib
python-jose[cryptography]
bcrypt==4.0.1
uvicorn
pydantic
openai
python-dotenv
tiktoken
PyPDF2
langchain
langchain-core
langchain-openai
langchain-community
langchain-chroma
chromadb
pydub
pydantic-settings
pytesseract
langgraph
aiohttp
aiofiles
deepgram-sdk
elevenlabs
psycopg2-binary
